requirements.txt
gunicorn
flask
psycopg2
requests
flask-cors
flask-sqlalchemy
